Refactor kvstore test (#13140)
* Refactor kvstore test

* Fix pylint

* Fix problem with some OSX not handling the cast on imDecode (#13207)

* Fix num_gpus
